ref(screens): Move module to performance dir
Moving the screens module out of the starfish directory and into the
performance directory for the long term
