+++ static/app/views/performance/screenload/screens/constants.ts
