+++ static/app/views/performance/screenload/screens/platformSelector.spec.tsx
@@ -3,7 +3,7 @@ import {browserHistory} from 'react-router';
 import {render, screen, userEvent} from 'sentry-test/reactTestingLibrary';
 
 import localStorage from 'sentry/utils/localStorage';
-import {PlatformSelector} from 'sentry/views/starfish/views/screens/platformSelector';
+import {PlatformSelector} from 'sentry/views/performance/screenload/screens/platformSelector';
 
 jest.mock('sentry/utils/localStorage');
