+++ static/app/views/performance/screenload/screens/screenBarChart.tsx
