+++ static/app/views/starfish/views/appStartup/screenSummary/eventSamples.tsx
@@ -8,6 +8,8 @@ import {decodeScalar, decodeSorts} from 'sentry/utils/queryString';
 import {MutableSearch} from 'sentry/utils/tokenizeSearch';
 import {useLocation} from 'sentry/utils/useLocation';
 import usePageFilters from 'sentry/utils/usePageFilters';
+import {EventSamplesTable} from 'sentry/views/performance/screenload/screenLoadSpans/eventSamplesTable';
+import {useTableQuery} from 'sentry/views/performance/screenload/screens/screensTable';
 import {
   PRIMARY_RELEASE_ALIAS,
   SECONDARY_RELEASE_ALIAS,
@@ -15,8 +17,6 @@ import {
 import {useReleaseSelection} from 'sentry/views/starfish/queries/useReleases';
 import {SpanMetricsField} from 'sentry/views/starfish/types';
 import {COLD_START_TYPE} from 'sentry/views/starfish/views/appStartup/screenSummary/startTypeSelector';
-import {EventSamplesTable} from 'sentry/views/starfish/views/screens/screenLoadSpans/eventSamplesTable';
-import {useTableQuery} from 'sentry/views/starfish/views/screens/screensTable';
 
 const DEFAULT_SORT: Sort = {
   kind: 'desc',
